+++ src/main/java/pl/multishop/service/ClientService.java
@@ -6,15 +6,10 @@ import java.util.List;
 
 public interface ClientService {
 
-    public Client findById(int clientId);
-
+    public Client findById(Long clientId);
     public Client findBySurname(String clientSurname);
-
     public void saveClient(Client client);
-
     public void updateClient(Client client);
-
     public void delClient(int clientId);
-
     public List<Client> findAllClients();
 }
